refactor:  Dropdown 컴포넌트 ts 파일로 마이그레이션
#1344 #1372
